safari-600.7-branch/LayoutTests/svg/animations/svgrect-animation-2-expected.txt

SVG 1.1 dynamic animation tests

Tests from-by SVGRect animation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rootSVGElement.viewBox.animVal.x is 0
PASS rootSVGElement.viewBox.animVal.y is 0
PASS rootSVGElement.viewBox.animVal.width is 100
PASS rootSVGElement.viewBox.animVal.height is 100
PASS rootSVGElement.viewBox.baseVal.x is 0
PASS rootSVGElement.viewBox.baseVal.y is 0
PASS rootSVGElement.viewBox.baseVal.width is 100
PASS rootSVGElement.viewBox.baseVal.height is 100
PASS rootSVGElement.viewBox.animVal.x is 25
PASS rootSVGElement.viewBox.animVal.y is 25
PASS rootSVGElement.viewBox.animVal.width is 125
PASS rootSVGElement.viewBox.animVal.height is 125
PASS rootSVGElement.viewBox.baseVal.x is 0
PASS rootSVGElement.viewBox.baseVal.y is 0
PASS rootSVGElement.viewBox.baseVal.width is 100
PASS rootSVGElement.viewBox.baseVal.height is 100
PASS rootSVGElement.viewBox.animVal.x is 50
PASS rootSVGElement.viewBox.animVal.y is 50
PASS rootSVGElement.viewBox.animVal.width is 150
PASS rootSVGElement.viewBox.animVal.height is 150
PASS rootSVGElement.viewBox.baseVal.x is 0
PASS rootSVGElement.viewBox.baseVal.y is 0
PASS rootSVGElement.viewBox.baseVal.width is 100
PASS rootSVGElement.viewBox.baseVal.height is 100
PASS rootSVGElement.viewBox.animVal.x is 0
PASS rootSVGElement.viewBox.animVal.y is 0
PASS rootSVGElement.viewBox.animVal.width is 100
PASS rootSVGElement.viewBox.animVal.height is 100
PASS rootSVGElement.viewBox.baseVal.x is 0
PASS rootSVGElement.viewBox.baseVal.y is 0
PASS rootSVGElement.viewBox.baseVal.width is 100
PASS rootSVGElement.viewBox.baseVal.height is 100
PASS successfullyParsed is true

TEST COMPLETE
